(fix) Moved the icons to its own component
