+++ packages/esm-patient-common-lib/src/dashboards/DashboardExtension.tsx
@@ -3,21 +3,8 @@ import classNames from 'classnames';
 import last from 'lodash-es/last';
 import { useTranslation } from 'react-i18next';
 import { useLocation } from 'react-router-dom';
-import {
-  ConfigurableLink,
-  ActivityIcon,
-  ShoppingCartIcon,
-  MedicationIcon,
-  ChartAverageIcon,
-  CalendarHeatMapIcon,
-  WarningIcon,
-  ListCheckedIcon,
-  DocumentAttachmentIcon,
-  EventScheduleIcon,
-  ReportIcon,
-  SyringeIcon,
-  ProgramsIcon,
-} from '@openmrs/esm-framework';
+import { ConfigurableLink } from '@openmrs/esm-framework';
+import { IconRenderer } from './MenuIcons';
 import styles from './dashboard-extension.scss';
 
 export interface DashboardExtensionProps {
@@ -25,50 +12,17 @@ export interface DashboardExtensionProps {
   title: string;
   basePath: string;
   moduleName?: string;
-  iconName?: string;
 }
 
-const MenuIcons = {
-  'Patient Summary': ReportIcon,
-  'Vitals & Biometrics': ActivityIcon,
-  Medications: MedicationIcon,
-  Orders: ShoppingCartIcon,
-  Results: ChartAverageIcon,
-  Visits: CalendarHeatMapIcon,
-  Allergies: WarningIcon,
-  Conditions: ListCheckedIcon,
-  Immunizations: SyringeIcon,
-  Attachments: DocumentAttachmentIcon,
-  Programs: ProgramsIcon,
-  Appointments: EventScheduleIcon,
-} as const;
-
-export type MenuTitle = keyof typeof MenuIcons;
-
-const DashboardExtension = ({
+export const DashboardExtension = ({
   path,
   title,
   basePath,
   moduleName = '@openmrs/esm-patient-chart-app',
-  iconName,
 }: DashboardExtensionProps) => {
   const { t } = useTranslation(moduleName);
   const location = useLocation();
   const navLink = useMemo(() => decodeURIComponent(last(location.pathname.split('/'))), [location.pathname]);
-
-  const renderIcon = () => {
-    if (iconName) {
-      const IconComponent = MenuIcons[title as MenuTitle];
-      return IconComponent ? <IconComponent className={styles.icon} /> : null;
-    }
-
-    if (title in MenuIcons) {
-      const IconComponent = MenuIcons[title as MenuTitle];
-      return <IconComponent className={styles.icon} />;
-    }
-
-    return null;
-  };
 
   return (
     <div key={path}>
@@ -77,12 +31,10 @@ const DashboardExtension = ({
         to={`${basePath}/${encodeURIComponent(path)}`}
       >
         <span className={styles.menu}>
-          {renderIcon()}
+          <IconRenderer title={title as any} />
           <span>{t(title)}</span>
         </span>
       </ConfigurableLink>
     </div>
   );
 };
-
-export default DashboardExtension;
