+++ packages/esm-patient-common-lib/src/dashboards/createDashboardLink.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { type DashboardLinkConfig } from '../types';
-import DashboardExtension from './DashboardExtension';
+import { DashboardExtension } from './DashboardExtension';
 
 export const createDashboardLink = (db: DashboardLinkConfig) => {
   return ({ basePath }: { basePath: string }) => {
